Using a configuration structure to aggregate all settable flags. Setting up module structure to be able to perform type checking. A separate module for parsing, and a sister package for type-checking WDL programs.
