+++ src/main/scala/wdlTools/syntax/AbstractSyntax.scala
@@ -1,18 +1,12 @@
-package wdlTools
+package wdlTools.syntax
 
-// A parser based on a WDL grammar written by Patrick Magee. The tool
-// underlying the grammar is Antlr4.
-//
-
-// A concrete syntax for the Workflow Description Language (WDL). This shouldn't be used
-// outside this package. Please use the abstract syntax instead.
-object ConcreteSyntax {
-  sealed trait Element
-  sealed trait WorkflowElement extends Element
-  sealed trait DocumentElement extends Element
+// An abstract syntax for the Workflow Description Language (WDL)
+object AbstractSyntax {
+  sealed trait WorkflowElement
+  sealed trait DocumentElement
 
   // type system
-  sealed trait Type extends Element
+  sealed trait Type
   case class TypeOptional(t: Type) extends Type
   case class TypeArray(t: Type, nonEmpty: Boolean) extends Type
   case class TypeMap(k: Type, v: Type) extends Type
@@ -27,38 +21,27 @@ object ConcreteSyntax {
   case class TypeStruct(name: String, members: Map[String, Type]) extends Type with DocumentElement
 
   // expressions
-  sealed trait Expr extends Element
-  case class ExprString(value: String) extends Expr
-  case class ExprFile(value: String) extends Expr
-  case class ExprBoolean(value: Boolean) extends Expr
-  case class ExprInt(value: Int) extends Expr
-  case class ExprFloat(value: Double) extends Expr
+  sealed trait Expr
+
+  // values
+  sealed trait Value extends Expr
+  case class ValueString(value: String) extends Value
+  case class ValueFile(value: String) extends Value
+  case class ValueBoolean(value: Boolean) extends Value
+  case class ValueInt(value: Int) extends Value
+  case class ValueFloat(value: Double) extends Value
 
   // represents strings with interpolation.
   // For example:
   //  "some string part ~{ident + ident} some string part after"
-  case class ExprCompoundString(value: Vector[Expr]) extends Expr
-  case class ExprMapLiteral(value: Map[Expr, Expr]) extends Expr
-  case class ExprObjectLiteral(value: Map[String, Expr]) extends Expr
-  case class ExprArrayLiteral(value: Vector[Expr]) extends Expr
-
   case class ExprIdentifier(id: String) extends Expr
+  case class ExprCompoundString(value: Vector[Expr]) extends Expr
+  case class ExprPair(l: Expr, r: Expr) extends Expr
+  case class ExprArray(value: Vector[Expr]) extends Expr
+  case class ExprMap(value: Map[Expr, Expr]) extends Expr
+  case class ExprObject(value: Map[String, Expr]) extends Expr
 
-  // These are parts of string interpolation expressions like:
-  //
-  // ${true="--yes" false="--no" boolean_value}
-  // ${default="foo" optional_value}
-  // ${sep=", " array_value}
-  //
-  trait PlaceHolderPart extends Element
-  // true="--yes"    false="--no"
-  case class ExprPlaceholderPartEqual(b: Boolean, value: Expr) extends PlaceHolderPart
-  // default="foo"
-  case class ExprPlaceholderPartDefault(value: Expr) extends PlaceHolderPart
-  // sep=", "
-  case class ExprPlaceholderPartSep(value: Expr) extends PlaceHolderPart
-
-  // These are full expressions of the same kind
+  // These are expressions of kind:
   //
   // ${true="--yes" false="--no" boolean_value}
   // ${default="foo" optional_value}
@@ -67,8 +50,11 @@ object ConcreteSyntax {
   case class ExprPlaceholderDefault(default: Expr, value: Expr) extends Expr
   case class ExprPlaceholderSep(sep: Expr, value: Expr) extends Expr
 
+  // operators on one argument
   case class ExprUniraryPlus(value: Expr) extends Expr
   case class ExprUniraryMinus(value: Expr) extends Expr
+
+  // operators on two arguments
   case class ExprLor(a: Expr, b: Expr) extends Expr
   case class ExprLand(a: Expr, b: Expr) extends Expr
   case class ExprNegate(value: Expr) extends Expr
@@ -83,17 +69,27 @@ object ConcreteSyntax {
   case class ExprMod(a: Expr, b: Expr) extends Expr
   case class ExprMul(a: Expr, b: Expr) extends Expr
   case class ExprDivide(a: Expr, b: Expr) extends Expr
-  case class ExprPair(l: Expr, r: Expr) extends Expr
+
+  // Access an array element at [index]
   case class ExprAt(array: Expr, index: Expr) extends Expr
-  case class ExprApply(funcName: String, elements: Vector[Expr]) extends Expr
+
+  // conditional:
+  // if (x == 1) then "Sunday" else "Weekday"
   case class ExprIfThenElse(cond: Expr, tBranch: Expr, fBranch: Expr) extends Expr
+
+  // Apply a standard library function to arguments. For example:
+  //   read_int("4")
+  case class ExprApply(funcName: String, elements: Vector[Expr]) extends Expr
+
+  // Access a field in a struct or an object. For example:
+  //   Int z = x.a
   case class ExprGetName(e: Expr, id: String) extends Expr
 
   case class Declaration(name: String, wdlType: Type, expr: Option[Expr]) extends WorkflowElement
 
   // sections
-  case class InputSection(declarations: Vector[Declaration]) extends Element
-  case class OutputSection(declarations: Vector[Declaration]) extends Element
+  case class InputSection(declarations: Vector[Declaration])
+  case class OutputSection(declarations: Vector[Declaration])
 
   // A command can be simple, with just one continuous string:
   //
@@ -108,24 +104,22 @@ object ConcreteSyntax {
   //     ls ~{input_file}
   //     echo ~{input_string}
   // >>>
-  case class CommandSection(parts: Vector[Expr]) extends Element
+  case class CommandSection(parts: Vector[Expr])
 
-  case class RuntimeKV(id: String, expr: Expr) extends Element
-  case class RuntimeSection(kvs: Vector[RuntimeKV]) extends Element
+  case class RuntimeKV(id: String, expr: Expr)
+  case class RuntimeSection(kvs: Vector[RuntimeKV])
 
   // meta section
-  case class MetaKV(id: String, expr: Expr) extends Element
-  case class ParameterMetaSection(kvs: Vector[MetaKV]) extends Element
-  case class MetaSection(kvs: Vector[MetaKV]) extends Element
+  case class MetaKV(id: String, expr: Expr)
+  case class ParameterMetaSection(kvs: Vector[MetaKV])
+  case class MetaSection(kvs: Vector[MetaKV])
 
-  // imports
-  case class ImportAlias(id1: String, id2: String) extends Element
-
-  // import statement as read from the document
-  case class ImportDoc(name: Option[String], aliases: Vector[ImportAlias], url: URL)
+  // import statement with the AST for the referenced document
+  case class ImportAlias(id1: String, id2: String)
+  case class ImportDoc(name: Option[String], aliases: Vector[ImportAlias], url: URL, doc: Document)
       extends DocumentElement
 
-  // top level definitions
+  // A task
   case class Task(name: String,
                   input: Option[InputSection],
                   output: Option[OutputSection],
@@ -136,26 +130,24 @@ object ConcreteSyntax {
                   runtime: Option[RuntimeSection])
       extends DocumentElement
 
-  case class CallAlias(name: String) extends Element
-  case class CallInput(name: String, expr: Expr) extends Element
-  case class CallInputs(value: Map[String, Expr]) extends Element
   case class Call(name: String, alias: Option[String], inputs: Map[String, Expr])
       extends WorkflowElement
+
   case class Scatter(identifier: String, expr: Expr, body: Vector[WorkflowElement])
       extends WorkflowElement
+
   case class Conditional(expr: Expr, body: Vector[WorkflowElement]) extends WorkflowElement
 
+  // A workflow
   case class Workflow(name: String,
                       input: Option[InputSection],
                       output: Option[OutputSection],
                       meta: Option[MetaSection],
                       parameterMeta: Option[ParameterMetaSection],
                       body: Vector[WorkflowElement])
-      extends Element
 
-  case class Version(value: String) extends Element
+  case class Version(value: String)
   case class Document(version: String,
                       elements: Vector[DocumentElement],
                       workflow: Option[Workflow])
-      extends Element
 }
