+++ src/main/scala/wdlTools/util/Util.scala
@@ -1,0 +1,11 @@
+package wdlTools.util
+
+import java.nio.file.{Path}
+
+object Util {
+
+  case class Conf(antlr4Trace: Boolean = false,
+                  verbose: Boolean = false,
+                  quiet: Boolean = false,
+                  localDirectories: Vector[Path] = Vector.empty)
+}
