+++ src/test/scala/wdlTools/syntax/AbstractSyntaxTest.scala
@@ -1,10 +1,10 @@
-package wdlTools
+package wdlTools.syntax
 
 import AbstractSyntax._
+import collection.JavaConverters._
 import java.nio.file.{Path, Paths, Files}
 import org.scalatest.{FlatSpec, Matchers}
-//import org.scalatest.Inside._
-import collection.JavaConverters._
+import wdlTools.util.Util.Conf
 
 class AbstractSyntaxTest extends FlatSpec with Matchers {
 
@@ -13,6 +13,8 @@ class AbstractSyntaxTest extends FlatSpec with Matchers {
     val p2: Path = Paths.get(getClass.getResource("/tasks").getPath)
     Vector(p1, p2)
   }
+  private lazy val conf = Conf(antlr4Trace = false,
+                               localDirectories = wdlSourceDirs)
 
   private def getWdlSource(dirname: String, fname: String): String = {
     val p: String = getClass.getResource(s"/${dirname}/${fname}").getPath
@@ -21,7 +23,7 @@ class AbstractSyntaxTest extends FlatSpec with Matchers {
   }
 
   it should "handle import statements" in {
-    val pa = new ParseAll(antlr4Trace = false, localDirectories = wdlSourceDirs)
+    val pa = new ParseAll(conf)
     val doc = pa.apply(getWdlSource("workflows", "imports.wdl"))
 
     doc.version shouldBe ("1.0")
