+++ src/test/scala/wdlTools/syntax/package.scala
@@ -1,4 +1,4 @@
-package wdlTools
+package wdlTools.syntax
 
 import org.scalatest.Tag
 object Edge extends Tag("edge")
